### 2018-06-20 #231 @mbjoseph
Submission: smapr
The smapr package discovers, downloads, and extracts global soil moisture data from the NASA SMAP mission, producing `raster` objects from HDF5 files.

```
Package: smapr
Title: Acquisition and Processing of NASA Soil Moisture Active-Passive (SMAP) Data
Version: 0.1.2
    person("Maxwell", "Joseph", email = "maxwell.b.joseph@colorado.edu",
            role = c("aut", "cre")),
    person("Matthew", "Oakley", email = "matthew.oakley@colorado.edu",
            role = "aut"),
    person("Zachary", "Schira", email = "zasc3143@colorado.edu",
            role = "aut")
    )
    rappdirs (>= 0.3.1),
    raster (>= 2.5),
    rhdf5 (>= 2.14),
    rvest,
    xml2
Maintainer: Maxwell Joseph <maxwell.b.joseph@colorado.edu>
    Facilitates programmatic access to NASA Soil Moisture Active
    Passive (SMAP) data with R. It includes functions to search for, acquire,
    and extract SMAP data.
    rgdal,
    roxygen2,
    utils

```

-   URL for the package (the development repository, not a stylized html page): https://github.com/earthlab/smapr

[e.g., "data extraction, because the package parses a scientific data file format"]

data retrieval and munging - smapr finds and downloads data, and extracts gridded spatial from HDF5 files with the correct projection information.

Soil scientists, ecologists, and people working in the remote sensing domain could use the package to quantify changes in soil moisture, detect droughts, generate covariate layers for species distribution models, and validate other soil moisture data products.

There is another smapr package, but it does not support the breadth of SMAP data products that this package does, and development seems to have ceased ~3 years ago: https://github.com/strongh/smapr. This smapr package has more thorough documentation and tests, allows extraction of multiple soil moisture data products, and has been on CRAN since 2016.

- [ x ] does not violate the Terms of Service of any service it interacts with.
- [ x ] has a CRAN and OSI accepted license.
- [ x ] contains a README with instructions for installing the development version.
- [ x ] includes documentation with examples for all functions.
- [ x ] contains a vignette with examples of its essential functions and uses.
- [ x ] has a test suite.
- [ x ] has continuous integration, including reporting of test coverage, using services such as Travis CI, Coveralls and/or CodeCov.
- [ x ] I agree to abide by [ROpenSci's Code of Conduct](https://github.com/ropensci/onboarding/blob/master/policies.md#code-of-conduct) during the review process and in maintaining my package should it be accepted.

- [ x ] Do you intend for this package to go on CRAN? (it's already on CRAN: https://cran.r-project.org/package=smapr)

- [ x ] Does `R CMD check` (or `devtools::check()`) succeed?  Paste and describe any errors or warnings:

- [ x ] Does the package conform to [rOpenSci packaging guidelines](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md)? Please describe any exceptions:

Anyone with experience in programmatically finding and getting data from a https server, and/or working with HDF5 raster data would be great - perhaps:

  - Jeff Hollister (jhollist)
  - Scott Chamberlain (sckott)
  - Marco Sciaini (marcosci)
  - Michael Sumner (mdsumner)
