### 2018-06-21 #231 @noamross
Thanks for you submission, @mbjoseph!  This fits well within our scope.

`goodpractice::gp()` checks are pretty clean!  Please add relevant links to your description and check that long code lines are neccessary:

```
── GP smapr ────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in
    general. 6% of code lines are covered by test cases.

    R/download_smap.R:32:NA
    R/download_smap.R:33:NA
    R/download_smap.R:34:NA
    R/download_smap.R:35:NA
    R/download_smap.R:36:NA
    ... and 292 more lines

  ✖ add a "URL" field to DESCRIPTION. It helps users find information
    about your package online. If your package does not have a
    homepage, add an URL to GitHub, or the CRAN package package page.
  ✖ add a "BugReports" field to DESCRIPTION, and point it to a bug
    tracker. Many online code hosting services provide bug trackers for
    free, https://github.com, https://gitlab.com, etc.
  ✖ avoid long code lines, it is bad for readability. Also, many people
    prefer editor windows that are about 80 characters wide. Try make
    your lines shorter than 80 characters

    tests/testthat/test-extract_smap.R:117:1
    tests/testthat/test-find_smap.R:10:1
    tests/testthat/test-find_smap.R:20:1

───────────────────────────────────────────────

```

One real spelling error from `spelling::spell_check_package()`

```
 WORD                 FOUND IN
availble             smapr-intro.Rmd:36
```

A note: There's a poorly documented CRAN feature for including bioconductor packages using a `biocViews:` field in DESCRIPTION, which you should add to ease installation for users.   See [here](https://bioinformatics.stackexchange.com/questions/3365/r-package-development-how-does-one-automatically-install-bioconductor-packages).

I'll go ahead and seek reviewers.
