### 2018-06-21 472b449
Addressing Noam's comments
See ropensci/onboarding#231

- tyop in vignette
- include biocViews for hdf5
- add URL and BugReports links
- fix long lines in tests
