+++ DESCRIPTION
@@ -12,6 +12,7 @@ Authors@R: c(
     )
 Depends:
     R (>= 3.2.5)
+biocViews: 
 Imports:
     httr (>= 1.1.0),
     rappdirs (>= 0.3.1),
@@ -35,3 +36,5 @@ Suggests:
     testthat,
     utils
 VignetteBuilder: knitr
+URL: https://github.com/earthlab/smapr
+BugReports: https://github.com/earthlab/smapr/issues
