### 2018-07-23 earthlab/smapr#36 @mbjoseph
Update installation instructions for hdf5
As pointed out by @marcosci (https://github.com/ropensci/onboarding/issues/231), the rhdf5 package no longer needs to be installed separately before smapr because we now have `biocViews` in the DESCRIPTION. So, we need to update the installation instructions in the README to reflect the streamlined installation process.
