### 2018-07-23 earthlab/smapr#37 @mbjoseph
Mention httr system deps in README
@marcosci pointed out that it might be useful to mention the system dependencies for the httr package in the installation instructions (https://github.com/ropensci/onboarding/issues/231). On Debian systems this is `libssl-dev` and `openssl@1.1` via brew for macOS.
