### 2018-07-23 earthlab/smapr#38 @mbjoseph
Add visualization of typical workflow to README
@marcosci points to a neat example here: https://github.com/socialcopsdev/rLandsat in https://github.com/ropensci/onboarding/issues/231

For smapr this is:

`find_smap` ==> `download_smap` ==> `list_smap` ==> `extract_smap`

I wonder whether there's a way to programmatically do this without needing to create a figure in a vector graphics editor...
