### 2018-07-23 earthlab/smapr#39 @mbjoseph
Detail types of SMAP data
@marcosci points out in https://github.com/ropensci/onboarding/issues/231 that it would be worth mentioning the types of data you can get from the SMAP mission. Could add these to the README:

- surface soil moisture
- root zone soil moisture
- freeze/thaw status
- temperature
etc.
