### 2018-07-23 earthlab/smapr#38 @marcosci
Tikz? :-P

Or just make a network, name the nodes accordingly and plot it?
