### 2018-07-23 earthlab/smapr#40 @mbjoseph
Move Docker instructions under "Installation" section
From @marcosci as part of https://github.com/ropensci/onboarding/issues/231 - the Docker instructions are most relevant as an alternative means of installation, so they should live in the "Installation" section.
