### 2018-07-23 earthlab/smapr#41 @mbjoseph
Function to set credentials
@marcosci pointed out a convenient way for users to provide credentials through a function like

```r
set_credentials <- function(ed_un, ed_pw){
  write(paste0("Sys.setenv(ed_un = \"",ed_un,"\", ed_pw = \"", ed_pw, "\")"),
        file=file.path(Sys.getenv("HOME"), ".Rprofile"),append=TRUE)
}
```

Want to add, but will need to think about a good name. I'd like to get `smap` into the name somewhere, to avoid having too general a name.
