### 2018-07-23 499d954
Adding codemeta file
As suggested by @marcosci in https://github.com/ropensci/onboarding/issues/231

This also adds a repo status badge as recommended by the codemetar package.
