+++ README.Rmd
@@ -9,6 +9,8 @@ output: github_document
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![](http://cranlogs.r-pkg.org/badges/grand-total/smapr)](http://cran.rstudio.com/web/packages/smapr/index.html) 
 [![](https://badges.ropensci.org/231_status.svg)](https://github.com/ropensci/onboarding/issues/231)
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](http://www.repostatus.org/badges/latest/active.svg)](http://www.repostatus.org/#active)
+
 
 ```{r, echo = FALSE}
 knitr::opts_chunk$set(
