### 2018-07-24 #231 @mbjoseph
Thanks @marcosci and @ldecicco-USGS for your reviews! This is all great feedback.
