### 2018-07-24 earthlab/smapr#42 @mbjoseph
Vignette expansion
Both @marcosci and @ldecicco-USGS came up with some great suggestions to improve the vignette (https://github.com/ropensci/onboarding/issues/231):

- [x] Add some more detailed post-processing examples to illustrate things like cropping, masking, or filtering the soil moisture data. Maybe include some more interesting visualizations.

- [x] Give some indication for how long `download_smap` takes. This will depend on the data type, number of files, and the users internet connection, but it's worth at least mentioning that this can be either fast or slow.

- [x] List the supported SMAP data types in the vignette (same table from README).
