### 2018-07-24 earthlab/smapr#43 @mbjoseph
Check for 401 HTTP responses (e.g., if user has wrong password)
@ldecicco-USGS encountered some esoteric error/not-so-great behavior when accidentally providing the wrong credentials (https://github.com/ropensci/onboarding/issues/231). As she points out, it would be much better to check for 401 responses and let users know when they're getting unauthorized 401 responses from the data server with a nice message and suggested fix.
