### 2018-07-24 earthlab/smapr#44 @mbjoseph
Add verbose argument to download_smap()
@ldecicco-USGS suggests adding some message that indicates when files are downloading so users know that things are happening (https://github.com/ropensci/onboarding/issues/231). This will be particularly useful if people are downloading many large files over a slow connection.

The easiest way to do this would be to add a `verbose` argument to `download_smap` that optionally prints which files are being downloaded when.
