### 2018-07-27 32a59b3
Adding set_smap_credentials and checking for 401
This commit adds a function set_smap_credentials() that users can use to
specify their NASA Earthdata credentials. It also implements checks for
401/unauthorized responses from the SMAP data server, which indicate
that the credentials used are not authorized (would be expected if users
enter the wrong username and/or password).

Big thanks to @marcosci and @ldecicco-USGS for suggesting this
functionality. This should solve #43 and #41.
