+++ NAMESPACE
@@ -4,6 +4,7 @@ export(download_smap)
 export(extract_smap)
 export(find_smap)
 export(list_smap)
+export(set_smap_credentials)
 importFrom(httr,GET)
 importFrom(httr,authenticate)
 importFrom(httr,write_disk)
