### 2018-07-27 5c1280f
Add skip_on_cran to tests that need authentication or modify the global .Renviron file
