### 2018-07-27 f30a390
Trying to skip or fix failing tests on Travis
Apparently it's not possible to modify the global .Renviroment file on Travis CI
