### 2018-07-27 4021f52
Ensuring that ~/.Renviron exists on travis
