+++ tests/testthat/test-set_smap_credentials.R
@@ -2,7 +2,6 @@ context("set_smap_credentials")
 
 test_that(".Renviron file is not modified when save = FALSE", {
   skip_on_cran()
-  skip_on_travis()
   renvironment_contents <- readLines(renvironment_path)
   creds <- get_creds(renvironment_path)
   
@@ -20,7 +19,6 @@ test_that(".Renviron file is not modified when save = FALSE", {
 
 test_that("Existing credentials raise an error when overwrite = FALSE", {
   skip_on_cran()
-  skip_on_travis()
   expect_error(set_smap_credentials("dummy_user", 
                                     "dummy_password", 
                                     save = TRUE, 
@@ -30,7 +28,6 @@ test_that("Existing credentials raise an error when overwrite = FALSE", {
 
 test_that("Existing credentials are overwritten when overwrite = TRUE", {
   skip_on_cran()
-  skip_on_travis()
   original_creds <- get_creds(renvironment_path)
   
   set_smap_credentials("user", 
