### 2018-07-27 c07fcb1
Ensure Travis ~/.Renviron has creds
From what I can gather Travis defines ed_un and ed_pw as environment variables just fine, so they can be accessed with `r Sys.getenv()`. This ensures that these environment variables are defined in ~/.Renviron, as would be expected on a typical installation.

Also adds a test for get_creds() to address coverage.
