### 2018-07-27 earthlab/smapr#37 @mbjoseph
Actually - I think it makes more sense not to do this, as these system deps are already well handled by the httr package. If a user tries to install it without the proper system deps, they will get a nice error message, e.g.,

```
------------------------ ANTICONF ERROR ---------------------------
Configuration failed because openssl was not found. Try installing:
 * deb: libssl-dev (Debian, Ubuntu, etc)
 * rpm: openssl-devel (Fedora, CentOS, RHEL)
 * csw: libssl_dev (Solaris)
 * brew: openssl@1.1 (Mac OSX)
If openssl is already installed, check that 'pkg-config' is in your
PATH and PKG_CONFIG_PATH contains a openssl.pc file. If pkg-config
is unavailable you can set INCLUDE_DIR and LIB_DIR manually via:
R CMD INSTALL --configure-vars='INCLUDE_DIR=... LIB_DIR=...'
--------------------------------------------------------------------
ERROR: configuration failed for package ‘openssl’
```

So, that should provide a better hint (and one that does not need to be maintained as part of the smapr package) over the long term, consistent with the idea that the R packages smapr depends on are responsible for handling their own system dependencies.
