### 2018-07-27 4e6b426
Clean up installation section of README
Solves #40 by moving Docker section under "Installation"

Solves #36 by removing rhdf5 installation instructions (no longer necessary)
