+++ README.Rmd
@@ -25,14 +25,6 @@ An R package for acquisition and processing of [NASA (Soil Moisture Active-Passi
 
 ## Installation
 
-### Dependencies
-
-To install smapr, you'll need to install the R package
-[rhdf5](https://www.bioconductor.org/packages/release/bioc/html/rhdf5.html), 
-a Bioconductor package that will not be installed automatically when installing 
-smapr. 
-For rhdf5 installation instructions see http://bioconductor.org/packages/release/bioc/html/rhdf5.html
-
 To install smapr from CRAN: 
 
 ```{r cran-installation, eval = FALSE}
@@ -45,6 +37,18 @@ To install the development version from GitHub:
 # install.packages("devtools")
 devtools::install_github("earthlab/smapr")
 ```
+
+#### Docker instructions (alternative to a local installation)
+
+If a local installation is not possible for some reason, we have made a Docker 
+image available with smapr and all its dependencies.
+
+```
+docker run -d -p 8787:8787 earthlab/smapr
+```
+
+In a web browser, navigate to localhost:8787 and log in with 
+username: rstudio, password: rstudio.
 
 
 ## Authentication
@@ -182,17 +186,6 @@ writeRaster(sm_raster, "wgs84_ft.tif")
 file.remove("wgs84_ft.tif")
 ```
 
-### Running smapr in Docker
-
-To avoid dependency heck, we have made a Docker image available with smapr 
-and all its dependencies.
-
-```
-docker run -d -p 8787:8787 earthlab/smapr
-```
-
-In a web browser, navigate to localhost:8787 and log in with 
-username: rstudio, password: rstudio.
 
 ## Meta
 
