### 2018-07-30 aa3e52d
Use set_smap_credentials() in documentation
