+++ DESCRIPTION
@@ -33,6 +33,7 @@ Suggests:
     rgdal,
     rmarkdown,
     roxygen2,
+    sp,
     testthat,
     utils
 VignetteBuilder: knitr
