### 2018-07-30 d0e2285
Detail SMAP products & workflow
This addresses #38 and #39 by adding more detail on the types of data available and providing a high level overview of the typical workflow (although I opted out of a fancy visualization to keep things simple).

Also, I found a little typo in the vignette that this fixes.
