+++ README.md
@@ -36,15 +36,17 @@ Access to the NASA SMAP data requires authentication through NASA's Earthdata po
 
 Once you have an account, you need to pass your Earthdata username (`ed_un`) and password (`ed_pw`) as environmental variables that can be read from within your R session. There are a couple of ways to do this:
 
--   (Recommended): Use `set_smap_credentials('yourusername', 'yourpasswd')`. This will save your credentials for later use by default, and optionally overwrite existing credentials if `overwrite = TRUE`.
+### Recommended approach
+
+Use `set_smap_credentials('yourusername', 'yourpasswd')`. This will save your credentials by default, overwriting existing credentials if `overwrite = TRUE`.
+
+#### Alternative approaches
 
 -   Use `Sys.setenv()` interactively in your R session to set your username and password (not including the `<` and `>`):
 
 ``` r
 Sys.setenv(ed_un = "<your username>", ed_pw = "<your password>")
 ```
-
--   Use `Sys.setenv()` in your `.Rprofile` to set those environmental variables every time you load R.
 
 -   Create a text file `.Renviron` in your home directory, which contains your username and password. If you don't know what your home directory is, execute `normalizePath("~/")` in the R console and it will be printed. Be sure to include a new line at the end of the file or R will fail silently when loading it.
 
@@ -53,12 +55,12 @@ Example `.Renviron file` (note the new line at the end!):
     ed_un=slkdjfsldkjfs
     ed_pw=dlfkjDD124^
 
-Once this file is created, restart your R session and you should now be able to access these environment variables (e.g., via `Sys.getenv(ed_un)`).
+Once this file is created, restart your R session and you should now be able to access these environment variables (e.g., via `Sys.getenv("ed_un")`).
 
 SMAP data products
 ==================
 
-Multiple SMAP data products are provided by the NSIDC, and these products vary in the amount of processing. Currently, smapr primarily supports level 3 and level 4 data products, which represent global daily composite and global three hourly modeled data products, respectively. NSIDC provides documentation for all SMAP data products on their [website](https://nsidc.org/data/smap/smap-data.html), and we provide a summary of data products supported by smapr below.
+Multiple SMAP data products are provided by the NSIDC, and these products vary in the amount of processing. Currently, smapr primarily supports level 3 and level 4 data products, which represent global daily composite and global three hourly modeled data products, respectively. There are a wide variety of data layers available in SMAP products, including surface soil moisture, root zone soil moisture, freeze/thaw status, surface temperature, vegetation water content, vegetation opacity, net ecosystem carbon exchange, soil temperature, and evapotranspiration. NSIDC provides documentation for all SMAP data products on their [website](https://nsidc.org/data/smap/smap-data.html), and we provide a summary of data products supported by smapr below.
 
 | Dataset id  | Description                                         | Resolution |
 |-------------|-----------------------------------------------------|------------|
@@ -72,9 +74,21 @@ Multiple SMAP data products are provided by the NSIDC, and these products vary i
[…]
