### 2018-07-30 earthlab/smapr#38 @mbjoseph
Opted for a simpler solution - text! https://github.com/earthlab/smapr/commit/d0e2285f4079e7e5013827d0fc352a1064a6287f#diff-5140cd39fa2897e01ffbfb8db0999940R126
