### 2018-08-01 earthlab/smapr#50 @mbjoseph
Use pkgdown
As suggested in https://github.com/ropensci/onboarding/issues/231
