+++ README.Rmd
@@ -16,7 +16,7 @@ output: github_document
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  fig.path = "inst/img/"
+  fig.path = "man/figures/"
 )
 ```
 
