+++ README.md
@@ -94,9 +94,9 @@ Given a data frame produced by `find_smap`, `download_smap` downloads the data o
 
 ``` r
 downloads <- download_smap(available_data)
-#> Downloading https://n5eil01u.ecs.nsidc.org/SMAP/SPL3SMAP.003/2015.05.25/SMAP_L3_SM_AP_20150525_R13080_001.h5Downloading https://n5eil01u.ecs.nsidc.org/SMAP/SPL3SMAP.003/2015.05.25/SMAP_L3_SM_AP_20150525_R13080_001.qaDownloading https://n5eil01u.ecs.nsidc.org/SMAP/SPL3SMAP.003/2015.05.25/SMAP_L3_SM_AP_20150525_R13080_001.h5.iso.xml
-#> Downloading https://n5eil01u.ecs.nsidc.org/SMAP/SPL3SMAP.003/2015.05.25/SMAP_L3_SM_AP_20150525_R13080_001.h5Downloading https://n5eil01u.ecs.nsidc.org/SMAP/SPL3SMAP.003/2015.05.25/SMAP_L3_SM_AP_20150525_R13080_001.qaDownloading https://n5eil01u.ecs.nsidc.org/SMAP/SPL3SMAP.003/2015.05.25/SMAP_L3_SM_AP_20150525_R13080_001.h5.iso.xml
-#> Downloading https://n5eil01u.ecs.nsidc.org/SMAP/SPL3SMAP.003/2015.05.25/SMAP_L3_SM_AP_20150525_R13080_001.h5Downloading https://n5eil01u.ecs.nsidc.org/SMAP/SPL3SMAP.003/2015.05.25/SMAP_L3_SM_AP_20150525_R13080_001.qaDownloading https://n5eil01u.ecs.nsidc.org/SMAP/SPL3SMAP.003/2015.05.25/SMAP_L3_SM_AP_20150525_R13080_001.h5.iso.xml
+#> Downloading https://n5eil01u.ecs.nsidc.org/SMAP/SPL3SMAP.003/2015.05.25/SMAP_L3_SM_AP_20150525_R13080_001.h5
+#> Downloading https://n5eil01u.ecs.nsidc.org/SMAP/SPL3SMAP.003/2015.05.25/SMAP_L3_SM_AP_20150525_R13080_001.qa
+#> Downloading https://n5eil01u.ecs.nsidc.org/SMAP/SPL3SMAP.003/2015.05.25/SMAP_L3_SM_AP_20150525_R13080_001.h5.iso.xml
 str(downloads)
 #> 'data.frame':    1 obs. of  4 variables:
 #>  $ name     : chr "SMAP_L3_SM_AP_20150525_R13080_001"
@@ -126,7 +126,7 @@ sm_raster <- extract_smap(downloads, "Soil_Moisture_Retrieval_Data/soil_moisture
 plot(sm_raster, main = "Level 3 soil moisture: May 25, 2015")
 ```
 
-<img src="inst/img/extract-data-1.png" style="display: block; margin: auto;" />
+<img src="man/figures/extract-data-1.png" style="display: block; margin: auto;" />
 
 The path "Soil\_Moisture\_Retrieval\_Data/soil\_moisture" was determined from the output of `list_smap(downloads, all = TRUE)`, which lists all of the data contained in SMAP data files.
 
