+++ .github/issue_template.md
@@ -1,0 +1,7 @@
+# Before posting
+
+The GitHub issue tracker is intended for bug reports and feature requests. 
+Do not post your NASA Earthdata username or password with your issue!
+
+When you post, please include a minimal reproducible example of the problem 
+and/or desired behavior, if applicable. 
