### 2018-08-01 410f390
add credential info to vignette
