### 2018-08-01 #231 @mbjoseph
Thanks again @marcosci, @ldecicco-USGS, and @noamross for your time and reviews. Here are the point-by-point changes:

# Review 1 from @marcosci

## 1. Installation

- Removed now defunct rhdf5 instructions https://github.com/earthlab/smapr/issues/36
- Decided against adding extra hints for httr system deps, after a fair bit of deliberation https://github.com/earthlab/smapr/issues/37

## 2. Documentation

- Added condensed text of high level workflow https://github.com/earthlab/smapr/issues/38
- Described types of SMAP data that are available https://github.com/earthlab/smapr/issues/39
- Moved Docker instructions to Installation section of README https://github.com/earthlab/smapr/issues/40
- Added pkgdown version of docs https://github.com/earthlab/smapr/issues/50
- Expanded vignette to included cropping and masking https://github.com/earthlab/smapr/issues/42

## 3. Code

- Added `smapr::set_smap_credentials()` function to make authentication easier https://github.com/earthlab/smapr/issues/41

## 4. Future proofing

Thanks for pointing out the stars package! I'll keep my eye on it as a potential replacement for rhdf5 and raster.

## 5. Smaller ToDos

- Added codemeta file https://github.com/earthlab/smapr/blob/master/codemeta.json
- Put URL of pkgdown docs in GitHub repo description
- Added issue template https://github.com/earthlab/smapr/blob/master/.github/issue_template.md

# Review 2 from @ldecicco-USGS

## Documentation

- Added more detail to README on types of data available from the SMAP mission https://github.com/earthlab/smapr/issues/39

> My eyes swept over the Bioconductor instructions, so at first, it took me a bit to figure out how to install the rhdf5 package. Could you add the code directly?

Shoot! Fortunately this shouldn't be needed anymore with the `biocViews()` edit that @noamross suggested.

- Added a section to vignette to point people to NASA's Earthdata portal to get a username and password, with instructions to use `set_smap_credentials` https://github.com/earthlab/smapr/commit/410f3906175fb479eb8bd92409789cb580050463

- Added a function to check for 401 errors and return a helpful error message https://github.com/earthlab/smapr/issues/43

- Added `verbose` argument to `download_smap` that prints when files are being downloaded https://github.com/earthlab/smapr/issues/44

- Added note in vignette that downloading data may take a while https://github.com/earthlab/smapr/issues/42

- Added dataset ID table to vignette (from README) https://github.com/earthlab/smapr/issues/42

- Added some example downstream applications to vignette including cropping/masking to a spatial region, aggregating, and working with the actual values from the rasters to compare surface vs. rootzone soil moisture https://github.com/earthlab/smapr/issues/42

- Added pointer to Details section of `find_smap` help page for users to see all dataset IDs https://github.com/earthlab/smapr/commit/3b68bf560a6ff545000deeb25135a5dbcc293bbd

> Do you have a connection to the data?

Not really - other that we work in the same city! The National Snow and Ice Data Center (which hosts the data) is aware of the package, however.
