### 2018-08-01 #231 @mbjoseph
Sorry - just accidentally hit "Close and comment"! The above summarizes the changes that were made. I really appreciate everyone's help with this - the package is much improved now thanks to these detailed (and friendly :smiley:) reviews!
