### 2018-08-03 9a6e31c
Rbuildignore .github and docs/ https://github.com/ropensci/onboarding/issues/231
