### 2018-08-03 a8f9198
Add ropensci footer to README
