+++ README.Rmd
@@ -211,3 +211,5 @@ after reading our contribution [guidelines](CONTRIBUTING.md), and the
 [Contributor Code of Conduct](CONDUCT.md). 
 * License: GPL-3
 * See `citation("smapr")` in R to cite this package in publications. 
+
+[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
