### 2018-08-03 acfe300
Point to ropensci org instead of earthlab
