+++ README.Rmd
@@ -2,9 +2,9 @@
 title: "smapr"
 output: github_document
 ---
-[![Build Status](https://travis-ci.org/earthlab/smapr.svg?branch=master)](https://travis-ci.org/earthlab/smapr)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/earthlab/smapr?branch=master&svg=true)](https://ci.appveyor.com/project/mbjoseph/smapr)
-[![codecov](https://codecov.io/gh/earthlab/smapr/branch/master/graph/badge.svg)](https://codecov.io/gh/earthlab/smapr)
+[![Build Status](https://travis-ci.org/ropensci/smapr.svg?branch=master)](https://travis-ci.org/ropensci/smapr)
+[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/smapr?branch=master&svg=true)](https://ci.appveyor.com/project/mbjoseph/smapr)
+[![codecov](https://codecov.io/gh/ropensci/smapr/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/smapr)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/smapr)](https://cran.r-project.org/package=smapr)
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![](http://cranlogs.r-pkg.org/badges/grand-total/smapr)](http://cran.rstudio.com/web/packages/smapr/index.html) 
@@ -35,7 +35,7 @@ To install the development version from GitHub:
 
 ```{r gh-installation, eval = FALSE}
 # install.packages("devtools")
-devtools::install_github("earthlab/smapr")
+devtools::install_github("ropensci/smapr")
 ```
 
 #### Docker instructions (alternative to a local installation)
@@ -206,7 +206,7 @@ file.remove("sm_raster.tif")
 
 ## Meta
 
-* Please [report any issues or bugs](https://github.com/earthlab/smapr/issues),
+* Please [report any issues or bugs](https://github.com/ropensci/smapr/issues),
 after reading our contribution [guidelines](CONTRIBUTING.md), and the 
 [Contributor Code of Conduct](CONDUCT.md). 
 * License: GPL-3
