### 2018-08-03 dbf3db6
update codemeta to ropensci from earthlab
