### 2018-08-03 #231 @mbjoseph
Great - thanks @noamross! I have updated the .Rbuildignore file (https://github.com/ropensci/smapr/commit/9a6e31c1774120dda193229ed94ef4dcbcac0257), transferred the repo to the ropensci org, added the footer (https://github.com/ropensci/smapr/commit/a8f9198ba70aac7a0e69ad12e096d702bdea2b5f), updated badges (https://github.com/ropensci/smapr/commit/acfe30056817bbb562d5e728dccc3aaab63e9b0b), and updated the Codemeta file (https://github.com/ropensci/smapr/commit/dbf3db6735300a516464022ab12ca23000bcfcad).

I'd love to add @ldecicco-USGS and @marcosci as reviewers if they are comfortable with it.

I'd also be happy to write a blog post about smapr, and I'll take a look at the onboarding section of the gitbook!
