### 2018-08-03 #231 @noamross
Excellent.   @stefaniebutland will be in touch about the blog schedule, in the meantime instructions for putting together post are here. https://github.com/ropensci/roweb2/blob/master/readme.md.  I'm closing the issue but feel free to carry on that conversation here.

I've made you admin of the repo over the the rOpenSci org.  Your other collaborators have write permissions.  Feel free to increase their permission levels if needed be.
