### 2018-09-06 #231 @stefaniebutland
@mbjoseph Are you still good with a Sep 15 draft for publication 2018-09-25?
